Write a function that searches for a value in an array of integers using the Linear search algorithm
        Prototype : int linear_search(int *array, size_t size, int value);
        Where array is a pointer to the first element of the array to search in
        size is the number of elements in array
        And value is the value to search for
        Your function must return the first index where value is located
        If value is not present in array or if array is NULL, your function must return -1
        Every time you compare a value in the array to the value you are searching, you have to print this value (see example below)

+++ 0x1E-search_algorithms/0-linear.c
@@ -2,19 +2,19 @@
 /**
  * linear_search - searches a value in an array of integer using Linear search
  * @array: pointer to the first element of the array to search in
- * @size: s the number of elements in array
+ * @size: the number of elements in array
  * @value: is the value to search for
  * Return: return the first index where value is located or -1
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	unsigned int i;
 
 	if (array == NULL || size == 0)
 		return (-1);
 	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%zu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%d] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
